CloudEagle/requirements.txt
fastapi==0.135.3
uvicorn[standard]==0.44.0
langgraph==1.1.6
langchain==1.2.15
langchain-core==1.2.29
langchain-google-genai==4.2.1
langchain-openai==0.2.14
langsmith>=0.2.5
google-genai==1.73.1
pydantic==2.13.0
pydantic-settings==2.13.1
httpx==0.28.1
tenacity==9.1.4
python-dotenv==1.2.2
pytest==9.0.3
pytest-asyncio==1.3.0
slowapi==0.1.9
cachetools==7.0.5
structlog==25.5.0
streamlit==1.56.0
requests==2.33.1
motor==3.6.0
tiktoken==0.8.0
openai==1.58.1
